Change Google timeMax timestamp to T23:59:59.999Z
It appears that the Harvest range query is inclusive of the final date, while the Google query is exclusive.

That is, for a given range `20140410..20140410`, Harvest returns all time entries for the day April 10th, while Google returns no entries.

This appears to be because the timestamp appended to timeMax is currently `T00:00:00.000Z` (the start of the day). Changing it to `T23:59:59.999Z` makes the Google and Harvest queries consistent — a range of one day will return events for that one day, and a range of multiple days will include events on the last day.